fixed locations of files and removed redundant tests
